# team-17c-main/user-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User Microservice API
  description: >
    The User Microservice allows users to perform different operations on their accounts, while specifying 
    their own user_id in a header named 'user_id' where needed. The root address of this API is '/user'.
    
    
    Certain operations are only executable by users which hold a certain role, for instance, the '/admin' paths are
    only executable by users which obtained the 'Admin' role. 
    The role hierarchy is (Regular) User < Author < Admin when comparing 
    by permissions. Regular users and authors are not able to view other users' credentials, in fact, they are 
    only allowed to view other users' profiles (which do not contain sensitive attributes). Nobody is allowed to view 
    their own nor other users' passwords through the endpoints, modification of one's password is possible.
    
    
    Admins are users too in our application.

  version: 1.0.0
servers:
  - url: http://localhost:8080/user
...
components:
  parameters:
    UserIDHeader: # try using this when a header is specified with the user email
      name: user_id
      in: header
      description: Your User ID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    User:
      required:
        - email
        - password
      type: object
      x-class-extra-annotation : "@javax.persistence.Entity" # share JPA entity name
      x-implements: ['nl.tudelft.sem.template.user.models.UserInterface']
      properties:
        user_id:
          description: the ID of the user
          type: string
          format: uuid
          x-field-extra-annotation: "@javax.persistence.Id"
        email:
          description: email of user
          type: string
          format: email
          pattern: "\\w+@\\w+\\.\\w{2,}"
        password:
          type: string
          format: password
#        profile:
#          $ref: "#/components/schemas/UserProfile"
#          type: object
#          format: UserProfile
#          x-field-extra-annotation: "@javax.persistence.OneToOne"
    UserWithoutPassword:
      required:
        - email
      type: object
      properties:
        user_id:
          description: the ID of the user
          type: string
          format: uuid
        email:
          description: email of user
          type: string
          format: email
          pattern: "\\w+@\\w+\\.\\w{2,}"
#        profile:
#          $ref: "#/components/schemas/UserProfile"
    UserAuth:
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    UserProfile:
      required:
        - username
      type: object
      x-class-extra-annotation : "@javax.persistence.Entity"
      properties:
        # needed for usage as follower/following and finding user profile by user id (to prevent
        # other users from seeing your email/password)
        user_id:
          description: the ID of the user
          type: string
          format: uuid
          x-field-extra-annotation: "@javax.persistence.Id @javax.persistence.GeneratedValue"

#          x-annotations:
#            - "@javax.persistence.Id"
#            - "@GeneratedValue(generator = \"uuid2\")"
#            - "@GenericGenerator(name = \"uuid2\", strategy = \"org.hibernate.id.UUIDGenerator\")"
#            - "@Column(name = \"userId\", columnDefinition = \"VARCHAR(255)\")"
        username:
          type: string
          pattern: "^\\w{2,20}$"
#        email:
#          type: string
#          format: email
        name:
          type: string
        privacy:
          type: integer
          enum:
            - 0
            - 1
          x-enum-varnames:
            - PUBLIC
            - PRIVATE
          description: >
            Privacy ID:
             * `0` - Public Profile
             * `1` - Private Profile
        role:
          type: integer
          enum:
            - 0
            - 1
            - 2
          x-enum-varnames:
            - USER
            - AUTHOR
            - ADMIN
          description: >
            Role ID:
             * `0` - User Role
             * `1` - Author Role
             * `2` - Admin Role
        state:
          type: integer
          enum:
            - 0
            - 1
            - 2
          x-enum-varnames:
            - ACTIVE
            - INACTIVE
            - BANNED
          default: 0  # Set the default value to ACTIVE
          description: >
            State ID:
             * `0` - Active
             * `1` - Inactive
             * `2` - Banned
        bio:
          type: string
        avatar:
          description: The profile picture
          type: string
        location:
          type: string
        favoriteGenres:
          description: array of ids of favorite genres
          type: array
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            type: string
        favoriteBook:
          description: id of favorite book
          type: string

        # for followers and following, what if 2 users follow each other
        # let's say user1 and user2, what if we look at the profile of user1, this will
        # show user2 in the following/followers list, then user2 will show user1 in the following/followers list
        # and so on... doesn't this implementation create infinite loops?
        # maybe we can just store the user id's, but that's just problems for the code itself
        # we can leave this object as it is right now, since it doesn't matter for the OpenAPI stuff
        followers:
          type: array
          description: UserProfiles of followers
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            description: the ID of the profile
            type: string
            format: uuid
        following:
          type: array
          description: UserProfiles of the "followed"
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            description: the ID of the profile
            type: string
            format: uuid
        friends:
          type: array
          description: UserProfiles of friends, friends = users who follow each other back
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            description: the ID of the profile
            type: string
            format: uuid
        bookshelves:
          description: array of id's of the bookshelves
          type: array
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            type: string
            format: UUID
...
paths:
  /account/create:
    post:
      tags:
        - User
      summary: Create a new user.
      description: Create a user account using your username, (unique) email address and password.
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuth'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
          description: Created
        '400':
          description: Bad Request
        '403':
          description: Email is already in use
  /account/view:
    get:
      tags:
        - User
      summary: View your account details, excluding your password.
      description: View the details of your own account.
      operationId: viewAccount
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithoutPassword'
        '400':
          description: Bad Request
        '405':
          description: Invalid input
  /account/modify:
    put:
      tags:
        - User
      summary: Update your account details.
      description: Update the details of your own account.
      operationId: updateAccount
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
              # do we need the normal user here, since we can also update the password?
      responses:
        '200':
          description: Updated User Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithoutPassword'
        '400':
          description: Bad Request
        '405':
          description: Invalid input
  /account/delete:
    delete:
      tags:
        - User
      summary: Delete your User Account.
      description: Delete your own account.
      operationId: deleteAccount
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      responses:
        '200':
          description: Deleted User
        '400':
          description: Bad Request
        '405':
          description: Invalid input
...